7238 FIX check_smtp: Float warning / critical thresholds can now be configured using the GUI
The active check check_smtp already supported floating point warning and critical
thresholds, but the GUI only accepted integral numbers.

+++ web/plugins/wato/active_checks.py
@@ -1468,11 +1468,11 @@ register_rule(group,
                          Tuple(
                              title = _("Expected response time"),
                              elements = [
-                                 Integer(
+                                 Float(
                                      title = _("Warning if above"),
                                      unit = _("sec")
                                  ),
-                                 Integer(
+                                 Float(
                                      title = _("Critical if above"),
                                      unit = _("sec")
                                  ),
